container: Moved lw_shared_container out of admin
(cherry picked from commit 2004a99ee7ebc9923260bffb1da2fc265d147f17)
